2.10.9

+++ adm/MessageListPage.php
@@ -3,7 +3,7 @@
 /**
  * @project XG Proyect
  * @version 2.10.x build 0000
- * @copyright Copyright (C) 2008 - 2012
+ * @copyright Copyright (C) 2008 - 2016
  */
 
 define('INSIDE'  , TRUE);
@@ -22,8 +22,8 @@ if ($Observation != 1) die(message ($lang['404_page']));
 	$DelDat     = ( !empty($_POST['deldat']) ) ? TRUE : FALSE;
 	$CurrPage   = ( !empty($_POST['curr'])   ) ? $_POST['curr'] : 1;
 	$Selected   = ( !empty($_POST['sele'])   ) ? $_POST['sele'] : 0;
-	$SelType    = $_POST['type'];
-	$SelPage    = $_POST['page'];
+	$SelType    = isset($_POST['type']) ? $_POST['type'] : '';
+	$SelPage    = isset($_POST['page']) ? $_POST['page'] : '';
 
 	$ViewPage = 1;
 	if ( $Selected != $SelType )
@@ -45,7 +45,7 @@ if ($Observation != 1) die(message ($lang['404_page']));
 			$ViewPage = 1;
 
 	}
-	elseif ($Next   == TRUE && $_POST['page'])
+	elseif ($Next   == TRUE && isset($_POST['page']) && $_POST['page'])
 	{
 		if ($Selected < 100)
 			$Mess      = doquery("SELECT COUNT(*) AS `max` FROM {{table}} WHERE `message_type` = '". $Selected ."';", 'messages', TRUE);
@@ -60,19 +60,22 @@ if ($Observation != 1) die(message ($lang['404_page']));
 			$ViewPage = $MaxPage;
 	}
 
-	if ($_POST['delsel'] && $_POST['sele'] >= 1 && $_POST['page'])
+	if (isset($_POST['delsel']) && $_POST['delsel'] && isset($_POST['sele']) && $_POST['sele'] >= 1 && isset($_POST['page']) && $_POST['page'])
 	{
 		if ($DelSel == TRUE)
 		{
+                    if (is_array($_POST['sele'])) {
+                        
 			foreach($_POST['sele'] as $MessId => $Value)
 			{
 				if ($Value = "on")
 					doquery ( "DELETE FROM {{table}} WHERE `message_id` = '". $MessId ."';", 'messages');
-			}
+			}   
+                    }
 		}
 	}
 
-	if ($_POST['deldat'] && $_POST['sele'] >= 1 && is_numeric($_POST['selday']) && is_numeric($_POST['selmonth']) && is_numeric($_POST['selyear'])
+	if (isset($_POST['deldat']) && $_POST['deldat'] && $_POST['sele'] >= 1 && is_numeric($_POST['selday']) && is_numeric($_POST['selmonth']) && is_numeric($_POST['selyear'])
 		&& $_POST['page'])
 	{
 		if ($DelDat == TRUE)
@@ -122,6 +125,8 @@ if ($Observation != 1) die(message ($lang['404_page']));
 	elseif ($Selected == 100)
 		$Messages            = doquery("SELECT * FROM {{table}} ORDER BY `message_time` DESC LIMIT ". $StartRec .",25;", 'messages');
 
+                $parse['mlst_data_rows'] = '';
+                
 		while ($row = mysql_fetch_assoc($Messages))
 		{
 			$OwnerData = doquery ("SELECT `username` FROM {{table}} WHERE `id` = '". $row['message_owner'] ."';", 'users',TRUE);
